requirements.txt
torch==2.6.0
Pillow>=10.0,<11
requests>=2.31,<3
numpy
dashscope>=1.17,<2
# Qwen + 本地模型（CPU）所需依赖
# 若需 GPU/CUDA，请按 README 中的说明选择对应的索引安装 torch/torchvision

# 本地模型
torch>=2.2.0
torchvision>=0.17.0

# 核心深度学习框架
torch>=1.12.0
torchvision>=0.13.0

# Web框架
flask>=2.0.0

# Django兼容性（如果后端需要）
django>=3.2.0
djangorestframework>=3.12.0
django-filter>=2.4.0
paddlehub
paddlepaddle



# 可选：性能优化
# gunicorn>=20.1.0  # 生产环境部署
# opencv-python>=4.5.0  # 如果需要额外的图像处理

# 开发工具
# pytest>=6.0.0  # 单元测试
# black>=22.0.0  # 代码格式化
